air flow project/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-postgres
apache-airflow-providers-http
pandas
SQLAlchemy
scikit-learn
xgboost
imblearn
matplotlib
pickle-mixin
dbt-core
dbt-postgres